Jadudm/remove migration db (#3063)
* Updates to remove the extra DB

* Removing c2g DB.

This is *mostly* done. There are issues around  the naming of the S3
bucket.

* Renaming data bucket for S3 work on migration

* Adding ignore for the data folder

Since it should never be checked in, it should have an ignore.

* Updating docs.

* This completes the removal of the migration db

It also settles on a consistent name for the container that will hold
the Census data while loading/operating on it.

* Resolving conflicts.

* Linting.

* This fixes linting, although this file is ancillary.

* Fixing spurious bandit flag.

* Updates test for config

This now reflects everything being in one DB.

* Linting.

* Fixed typo

* Moving the historical migration bucket to preview

This puts it in the preview space ONLY.

* terraform fmt

* Attempting to add the bucket to preview only

Broke out the list of services to be dependent on the environment.

* Attempting to include variables...

This is attempting to concat lists in YAML.

* For testing manifest. TEMP.

To be undone.

FIXME.

* remove unused var

* terraform fmt

* replace org name

* In order to bind...

Testing on preview ithout pushing to `main`.

* Test additional_services syntax

* test anchor

* Move anchor into main manifest file

* add anchor to history so we don't lose it

* Remove additional_services anchor

* Reset Manifest

* Use a seperate manifest for preview

* New preview deploy file

* change command

* change file name

* Make file executable

* Use actual bucket name

* Re-Enable Testing

* Code improvement

* Removed fac-census-data-s3 service

* Bug fix

* Shortcutting preview deploy for test purpose

---------

Co-authored-by: Hassan D. M. Sambo <hassandeme.mamasambo@gsa.gov>
Co-authored-by: Alex Steel <130377221+asteel-gsa@users.noreply.github.com>