Media Backups Patch (#2820)
* Expand the proxy for media backups

The following need to be added to our allow list to ensure that we can
get aws-cli and aws-s3-tar-tool onto the instance

* Remove the unset proxy

* Go through the proxy for the s3 tar binary

* Add notes for backups

This is temporary and will be removed at some point

* Fix Workflow

* Update notes

* Fix notational command to extract

* Move to a more permenant location

* Fix list contents

* Change restart command for "rolling" restart

* Remove if conditional on dev for media backups

* Simplifying the backup script

We discovered we can tar directly to another bucket. This allows us to
tar the contents of the media bucket directly to the backups bucket with
no intermediate files.

* Snapshot and sync

We now have two scripts. One syncs, and another to to a tar snapshot.
The latter is the original tar script, tweaked for smallification.

* Remove original backup workflow

* Add media sync workflow

* add tar snapshot workflow

* Scheduled dev backups

* scheduled prod backups

* Renaming

* update docs

---------

Co-authored-by: Matt Jadud <matthew.jadud@gsa.gov>